+++ R/elo.players.R
@@ -45,3 +45,30 @@ weights.elo.players.matrix <- function(object, ...) attr(object, "weights")
 
 is.players <- function(x) inherits(x, "elo.players.matrix")
 
+#' @rdname formula.specials
+#' @export
+multiteam <- function(...)
+{
+  args <- lapply(list(...), as.character)
+  as.elo.multiteam.matrix(do.call(cbind, args))
+}
+
+as.elo.multiteam.matrix <- function(x)
+{
+  if(!is.matrix(x)) stop("'x' isn't a matrix.")
+  if(nrow(x)*ncol(x) == 0) stop('No rows or no columns.')
+  structure(x, class = c("elo.multiteam.matrix", class(x)))
+}
+
+#' @export
+`[.elo.multiteam.matrix` <- function(x, i, j, drop = FALSE)
+{
+  if(!missing(j)) return(NextMethod())
+  as.elo.multiteam.matrix(unclass(x)[i, , drop = FALSE])
+}
+
+#' @export
+length.elo.multiteam.matrix <- function(x) nrow(x)
+
+#' @export
+is.na.elo.multiteam.matrix <- function(x) rowSums(is.na(unclass(x))) == ncol(unclass(x))